Added UI logic and UI to show Fibonacci list, loading and empty state

+++ app/src/main/java/com/flourenco/fibonacci/core/storage/database/entity/CalculatedFibonacciEntity.kt
@@ -7,5 +7,5 @@ import org.threeten.bp.LocalDateTime
 @Entity
 data class CalculatedFibonacciEntity(
     @PrimaryKey val calculatedNumber: Int,
-    val fibonacciValue: Int
+    val fibonacciValue: Long
 )
